NO-REF: First pass - move out extraneous kmeans logic
