+++ etl-pipeline/tests/unit/test_kmeans_manager.py
@@ -1,7 +1,8 @@
 import pandas as pd
 import pytest
 
-from managers.kMeans import KMeansManager, YearObject
+from managers.kMeans import KMeansManager
+from parsers import YearParser, get_publication_date_object
 
 
 class TestKMeansModel(object):
@@ -230,48 +231,49 @@ class TestKMeansModel(object):
         assert testModel.maxK == int(350 * (4 / 9))
 
     def test_getPubDateObject_range(self, mocker, TestYear):
-        mockParser = mocker.patch("managers.kMeans.YearObject")
+        mockParser = mocker.patch("parsers.date_parser.YearParser")
         mockParser.side_effect = [TestYear(20, 0, 0), TestYear(19, 0, 0)]
 
         mockDates = ["2000|other_date", "1900-1901|publication_date"]
-        outYear = KMeansManager.getPubDateObject(mockDates)
+        outYear = get_publication_date_object(mockDates)
 
         assert outYear == {"century": 19, "decade": 0, "year": 0}
 
     def test_getPubDateObject_year_only(self, mocker, TestYear):
-        mockParser = mocker.patch("managers.kMeans.YearObject")
+        mockParser = mocker.patch("parsers.date_parser.YearParser")
         mockParser.side_effect = [TestYear(20, 0, 0), TestYear(19, 0, 0)]
 
         mockDates = ["2000|other_date", "1900|publication_date"]
-        outYear = KMeansManager.getPubDateObject(mockDates)
+        outYear = get_publication_date_object(mockDates)
 
         assert outYear == {"century": 19, "decade": 0, "year": 0}
 
     def test_getPubDateObject_full_date(self, mocker, TestYear):
-        mockParser = mocker.patch("managers.kMeans.YearObject")
+        mockParser = mocker.patch("parsers.date_parser.YearParser")
         mockParser.side_effect = [TestYear(20, 0, 0), TestYear(19, 0, 0)]
 
         mockDates = ["2000-12-01|other_date", "1900-01-01|publication_date"]
-        outYear = KMeansManager.getPubDateObject(mockDates)
+        outYear = get_publication_date_object(mockDates)
+        print(outYear)
 
         assert outYear == {"century": 19, "decade": 0, "year": 0}
 
     def test_getPubDateFloat_neither(self, mocker):
-        mocker.patch("managers.kMeans.YearObject")
+        mocker.patch("parsers.date_parser.YearParser")
         mockDates = ["2000|other_date", "1900-1901|pub_date"]
-        outYear = KMeansManager.getPubDateObject(mockDates)
+        outYear = get_publication_date_object(mockDates)
         assert outYear == {}
 
     def test_getPubDateFloat_no_dates(self, mocker):
-        outYear = KMeansManager.getPubDateObject([])
+        outYear = get_publication_date_object([])
         assert outYear == {}
 
     def test_getPubDateFloat_bad_value(self, mocker):
-        outYear = KMeansManager.getPubDateObject(["2000"])
+        outYear = get_publication_date_object(["2000"])
         assert outYear == {}
 
     def test_getPubDateFloat_bad_string_value(self, mocker):
-        outYear = KMeansManager.getPubDateObject(["test|date"])
+        outYear = get_publication_date_object(["test|date"])
         assert outYear == {}
 
     def test_getPublishers_clean(self, mocker):
@@ -361,7 +363,7 @@ class TestKMeansModel(object):
         mockPipeline.fit_predict.assert_called_once()
 
     def test_parseEditions(self, testModel, mocker):
-        mockConvert = mocker.patch.object(YearObject, "convertYearDictToStr")
+        mockConvert = mocker.patch.object(YearParser, "convertYearDictToStr")
         mockConvert.side_effect = [1900, 1900, 2000, 1950]
         outEditions = testModel.parseEditions()
         assert len(outEditions) == 3
